Try to mitigate #30
As it is hard to reproduce the bug (race condition), we try to
mitigate. Here we assume the cause of the bug is the fact that a
PA scale update is coming before the status icon is completely
intialized, returning meaningful positional information which the
sliders window depends on.

If this assumption turns out to be wrong, this commit should
probably be reverted.